refactor(app): organize components by feature (issue #55)
- Move authentication-specific components to features/authentication.
- Move home-specific components to features/home.
- Move user-specific components to features/users.
- Keep generic atoms, pagination, and navigation shell in shared components.
- Update all internal and external import paths in views and tests.
- Ensure all 230 tests pass and analysis is clean.
