+++ app/lib/features/authentication/presentation/components/templates/auth_template.dart
