+++ app/lib/features/authentication/presentation/views/signin_screen.dart
@@ -3,8 +3,8 @@ import 'package:flutter/services.dart';
 import 'package:go_router/go_router.dart';
 import 'package:gatuno/l10n/app_localizations.dart';
 import '../../../../shared/icons/icons.dart';
-import '../../../../shared/components/organisms/signin_form.dart';
-import '../../../../shared/components/templates/auth_template.dart';
+import '../components/organisms/signin_form.dart';
+import '../components/templates/auth_template.dart';
 import '../../domain/use_cases/auth_service.dart';
 import '../view_models/signin_view_model.dart';
 import '../../../../core/di/injection.dart';
