+++ app/lib/features/home/presentation/components/templates/home_template.dart
