+++ app/lib/features/users/presentation/views/me_screen.dart
@@ -3,9 +3,9 @@ import 'package:provider/provider.dart';
 import 'package:go_router/go_router.dart';
 import 'package:gatuno/l10n/app_localizations.dart';
 import '../view_models/me_view_model.dart';
-import '../../../../shared/components/molecules/user_profile_header.dart';
+import '../components/molecules/user_profile_header.dart';
 import 'package:gatuno/features/settings/presentation/components/organisms/me_settings_list.dart';
-import '../../../../shared/components/templates/profile_template.dart';
+import '../components/templates/profile_template.dart';
 
 class MePage extends StatefulWidget {
   const MePage({super.key});
