+++ app/test/shared/components/organisms/signin_form_test.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:gatuno/shared/components/organisms/signin_form.dart';
+import 'package:gatuno/features/authentication/presentation/components/organisms/signin_form.dart';
 import 'package:gatuno/shared/components/atoms/app_button.dart';
 import 'package:gatuno/shared/components/atoms/app_text_field.dart';
 import '../../../helpers/pump_app.dart';
